Funcion de editar usuario dinamica y validacion para editar
