+++ NotasEquipo4/src/app/users-update/users-update.component.html
@@ -13,11 +13,11 @@
     </thead>
     <tbody>
         <tr *ngFor= "let userRegister of user">
-            <td><input type="text" name="nombre" value={{userRegister.nombre}}></td>
-            <td><input type="text" name="apellidos" value={{userRegister.apellidos}}></td>
-            <td><input type="text" name="telefono" value={{userRegister.telefono}}></td>
+            <td><input type="text" name="nombre" id="nombre"value={{userRegister.nombre}}></td>
+            <td><input type="text" name="apellidos" id="apellidos" value={{userRegister.apellidos}}></td>
+            <td><input type="text" name="telefono" id="telefono" value={{userRegister.telefono}}></td>
             <td>{{userRegister.email}}</td>
-            <td><input class="btn btn-primary" type="submit" value="Guardar"></td>
+            <td><input class="btn btn-primary" type="submit" (click)="editarUsuario(userRegister)" value="Guardar"></td>
        </tr>
     </tbody>
