+++ NotasEquipo4/src/app/users/users.component.html
@@ -19,7 +19,7 @@
             <td>{{userRegister.telefono}}</td>
             <td>{{userRegister.email}}</td>
             <td><button class="btn btn-danger" (click)="eliminarUsuario(userRegister)">Eliminar</button>
-                <button class="btn btn-primary" name="button" [routerLink]="['/users-update',userRegister.email]">Modificar</button></td>
+                <button class="btn btn-primary" name="button" (click)="esValido(userRegister)">Modificar</button></td>
         </tr>
     </tbody>
