+++ NotasEquipo4/src/app/users/users.component.ts
@@ -1,4 +1,6 @@
+import { Route } from '@angular/compiler/src/core';
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 import { Note } from '../notes/notes.module';
 import { Register} from '../register/registers.module';
 import { UserService } from '../services/user.service';
@@ -11,7 +13,7 @@ export class UsersComponent implements OnInit {
  // @Input() users: User[] = [];
 usuarios: Register[] =[];
 notes: Note[]=[];
-  constructor(private userService: UserService) { }
+  constructor(private userService: UserService, private router: Router) { }
 
   ngOnInit(): void {
     this.getUsers();
@@ -56,25 +58,14 @@ notes: Note[]=[];
     }
   }
 
-
-   //METODO PARA EDITAR EL USUARIO DE MANERA ESTATICA
-  public editarUsuario(usuarios : Register){
-    if(usuarios.email == localStorage.emailLogged){
-      let usersArray = JSON.parse(localStorage.users);
-      var nuevoTelefono = "1234567890"
-      usersArray.forEach(function(item:any) {
-        if(item.email == localStorage.emailLogged)
-          item.telefono = nuevoTelefono;
-          localStorage.setItem("users", JSON.stringify(usersArray));
-      });
-      console.log(usersArray)
-      //this.router.navigate(["new-notes"],{queryParams: {back_url: this.router.url}})
+  esValido(usuario:Register){
+    if(usuario.email == localStorage.emailLogged){
+      this.router.navigate(["/users-update","usuario.email"],{queryParams: {back_url: this.router.url}})
     }
     else{
-      alert("No puedes modificar este usuario")
+      alert("No coinciden los correos")
     }
   }
-
 
 
 }
